Base Gun
Adding in a Base Gun, it has all the variables and events it needs so we can do childs for each weapon we add in. At the moment I don't know if the events needs to be Enhanced Inputs or Not, and will need to ask Brian if I can do that to the project settings.