Adding in a base weapon
Added a base weapon blueprint named 'gun' with the basic variables and enhanced actions for shooting, aiming and reloading for others to make a child off of to make the other weapons quicker and easily. But the base blueprint doesn't have any special code in it. It's just the Enhanced Actions and those are left blank to be bettered edited based on the weapon.