Added explosive barrels
Barrels have a radius, amount of damage, and a fall off percent.

On explosion, check each object in the radius for the damage interface.  If so, then calculate the object's distance away from the center of the explosion and if the percent of the distance compared to the radius is below the fall off percent, then fall off damage is applied compared to the distance from the point where fall off begins.

That is my best explanation of how it works, hope it makes sense.